5.14 - js refactoring
