+++ js/utils.js
@@ -1,0 +1,3 @@
+const getRandomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+export default getRandomInteger;
